Move all logic out of index.ts to tasks and utils folders/files
